requirements.txt
asgiref==3.5.2
certifi==2022.12.7
charset-normalizer==2.1.1
Django==3.2
django-elasticsearch-dsl==7.2.2
django-elasticsearch-dsl-drf==0.22.5
django-nine==0.2.7
djangorestframework==3.14.0
docopt==0.6.2
elasticsearch==7.17.8
elasticsearch-dsl==7.4.0
idna==3.4
packaging==22.0
pipreqs==0.4.11
python-dateutil==2.8.2
pytz==2022.6
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.13
yarg==0.1.9
